adding two new libraries
the sector library works with the Rusko version of the light suit and
allows specific sections to be written on or off. It had memory
problems when many sectors were instantiated.

the memoryFree library was found on google. It is a test code to check
the amount of memory free on an Arduino processor