adding 4 libraries
GAY
4 LIBRARIES

Grid - for working with any size array of LPD8806 strips and contains
initialization procedures as well as functions to control the grid's
behavior.
ColorList -  defines 24-bit colors or some shit like that. the LPD
library has a function that can read in this number and spit out a
color onto a pixel. ***Doesn't work that well for some colors. Needs
de-bugging.
LEDArray - same as Grid, but was the original version and is only
designed to work with this gay bullshit 56-LED array Ryan made a while
ago on shitty cardboard.
LPD8806 - the library files for the strips. Haven't been edited at all.